# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentIntelligenceInput": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "document_refs": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "minItems": 0,
            "title": "Document Refs",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "document_refs"
        ],
        "title": "DocumentIntelligenceInput",
        "type": "object"
      },
      "DocumentIntelligenceOutput": {
        "description": "Aggregate of ExtractedFields across all input document_refs.\n\nNote (Story 3.4 \u00a7 AC8): ``evidence_ids`` inside each ``ExtractedField.value``\nare empty when the agent returns. The supervisor (Story 3-5) back-fills\nthem with the agent's own ledger-entry ID after ``@agent_action`` writes\nthe entry.",
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "extracted_fields": {
            "items": {
              "$ref": "#/components/schemas/ExtractedField"
            },
            "title": "Extracted Fields",
            "type": "array"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "DocumentIntelligenceOutput",
        "type": "object"
      },
      "DraftedRationale": {
        "description": "The Writing agent's output \u2014 a structured rationale with\ncitations.\n\n``html`` is the renderable form for Tiptap (citation tokens already\nwrapped in ``<span data-ledger-id=\"\u2026\">\u2026</span>``). ``paragraphs``\nand ``cited_claims`` carry the structured signal for downstream\nanalytics \u2014 they are persisted on the intake row but not surfaced\nin the demo UI.",
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "cited_claims": {
            "items": {
              "$ref": "#/components/schemas/CitedClaim"
            },
            "title": "Cited Claims",
            "type": "array"
          },
          "html": {
            "minLength": 20,
            "title": "Html",
            "type": "string"
          },
          "model_id": {
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "paragraphs": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 2,
            "title": "Paragraphs",
            "type": "array"
          },
          "prompt_template_id": {
            "const": "rationale_draft_v1",
            "default": "rationale_draft_v1",
            "title": "Prompt Template Id",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "html",
          "paragraphs",
          "model_id"
        ],
        "title": "DraftedRationale",
        "type": "object"
      },
      "EntityVerificationInput": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "cin": {
            "maxLength": 21,
            "minLength": 21,
            "pattern": "^[LU]\\d{5}[A-Z]{2}\\d{4}[A-Z]{3}\\d{6}$",
            "title": "Cin",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "cin"
        ],
        "title": "EntityVerificationInput",
        "type": "object"
      },
      "EntityVerificationResult": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "cin": {
            "title": "Cin",
            "type": "string"
          },
          "mca_status": {
            "$ref": "#/components/schemas/ProvenancedField_Literal__active____struck_off____dormant___"
          },
          "mismatches": {
            "items": {
              "$ref": "#/components/schemas/FieldMismatch"
            },
            "title": "Mismatches",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "cin",
          "mca_status"
        ],
        "title": "EntityVerificationResult",
        "type": "object"
      },
...
      "ExtractedField": {
        "properties": {
          "document_ref": {
            "minLength": 1,
            "title": "Document Ref",
            "type": "string"
          },
          "field_name": {
            "minLength": 1,
            "title": "Field Name",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/ProvenancedField_Union_str__int__float__bool__NoneType__"
          }
        },
        "required": [
          "field_name",
          "document_ref",
          "value"
        ],
        "title": "ExtractedField",
        "type": "object"
      },
      "FieldMismatch": {
        "properties": {
          "case_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Value"
          },
          "field_name": {
            "minLength": 1,
            "title": "Field Name",
            "type": "string"
          },
          "mca_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mca Value"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "severity": {
            "default": "warning",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "field_name"
        ],
        "title": "FieldMismatch",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvenancedField_Literal__active____struck_off____dormant___": {
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "value": {
            "enum": [
              "active",
              "struck_off",
              "dormant"
            ],
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "provenance"
        ],
        "title": "ProvenancedField[Literal['active', 'struck_off', 'dormant']]",
        "type": "object"
      },
...
      "ProvenancedField_float_": {
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value",
          "provenance"
        ],
        "title": "ProvenancedField[float]",
        "type": "object"
      },
...
      "RiskComponent": {
        "properties": {
          "contribution": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Contribution",
            "type": "number"
          },
          "name": {
            "enum": [
              "country",
              "entity_type",
              "ownership_clarity",
              "screening",
              "adverse_media"
            ],
            "title": "Name",
            "type": "string"
          },
          "rationale": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "value": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "number"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "name",
          "value",
          "weight",
          "contribution",
          "rationale"
        ],
        "title": "RiskComponent",
        "type": "object"
      },
      "RiskScore": {
        "properties": {
          "band": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Band",
            "type": "string"
          },
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/RiskComponent"
            },
            "title": "Components",
            "type": "array"
          },
          "score_provenance": {
            "$ref": "#/components/schemas/ProvenancedField_float_"
          },
          "total": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "case_id",
          "total",
          "band",
          "score_provenance"
        ],
        "title": "RiskScore",
        "type": "object"
      },
      "RiskScoringInput": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "RiskScoringInput",
        "type": "object"
      },
...
      "ScreeningAgentInput": {
        "description": "Input shape for the Story 6.2 Screening agent.\n\nSubjects are built by the supervisor from upstream agent outputs\n(Entity Verification + UBO Graph) \u2014 the agent receives them ready-built.",
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/ScreeningSubject"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Subjects",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "subjects"
        ],
        "title": "ScreeningAgentInput",
        "type": "object"
      },
      "ScreeningAgentOutput": {
        "description": "Output of the Screening agent.\n\n``hits`` includes auto-dismissed entries \u2014 the UI shows the dismissed\nones in a collapsed group so officers can re-include if needed.",
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/ScreeningHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "subjects_screened": {
            "minimum": 1.0,
            "title": "Subjects Screened",
            "type": "integer"
          }
        },
        "required": [
          "case_id",
          "subjects_screened"
        ],
        "title": "ScreeningAgentOutput",
        "type": "object"
      },
...
      "ScreeningSubject": {
        "description": "A single name+DOB+identifier triple to screen.",
        "properties": {
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "identifiers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Identifiers",
            "type": "object"
          },
          "subject_id": {
            "minLength": 1,
            "title": "Subject Id",
            "type": "string"
          },
          "subject_kind": {
            "enum": [
              "entity",
              "director",
              "ubo"
            ],
            "title": "Subject Kind",
            "type": "string"
          }
        },
        "required": [
          "subject_kind",
          "subject_id",
          "full_name"
        ],
        "title": "ScreeningSubject",
        "type": "object"
      },
...
      "UBOGraph": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/UBOEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UBOPersonNode"
                },
                {
                  "$ref": "#/components/schemas/UBOEntityNode"
                }
              ]
            },
            "title": "Nodes",
            "type": "array"
          },
          "root_entity_id": {
            "maxLength": 72,
            "minLength": 7,
            "pattern": "^ubo_(p|e)_[A-Za-z0-9_-]{1,64}$",
            "title": "Root Entity Id",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "root_entity_id"
        ],
        "title": "UBOGraph",
        "type": "object"
      },
      "UBOGraphInput": {
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          },
          "cin": {
            "maxLength": 21,
            "minLength": 21,
            "title": "Cin",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "cin"
        ],
        "title": "UBOGraphInput",
        "type": "object"
      },
      "UBOPersonNode": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "din": {
            "anyOf": [
              {
                "pattern": "^\\d{8}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Din"
          },
          "id": {
            "maxLength": 72,
            "minLength": 7,
            "pattern": "^ubo_(p|e)_[A-Za-z0-9_-]{1,64}$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "person",
            "default": "person",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UBOPersonNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WritingAgentInput": {
        "description": "Tool-facing input for the Writing agent. Upstream typed outputs\nare NOT in the input \u2014 the supervisor reads them off the case's\nintake row at call time, mirroring Story 6.2's screening agent.",
        "properties": {
          "case_id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "WritingAgentInput",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents/document_intelligence/extract": {
      "post": {
        "description": "Calls the Document Intelligence agent with the supplied case id and document refs. Returns the extracted fields with provenance + confidence band. Every invocation writes one ledger entry.",
        "operationId": "extract_document_fields_v1_agents_document_intelligence_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentIntelligenceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIntelligenceOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the Document Intelligence agent against a case's documents",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/entity_verification/verify": {
      "post": {
        "description": "Calls the Entity Verification agent. Looks the CIN up against the mock MCA lookup tool, diffs case-side fields, returns typed mismatches. Every invocation writes one ledger entry.",
        "operationId": "verify_entity_v1_agents_entity_verification_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityVerificationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityVerificationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the Entity Verification agent against a CIN",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/risk_scoring/score": {
      "post": {
        "description": "Calls the Risk Scoring agent. Reads prior intake outputs (entity_verification, ubo_graph) and customer_metadata to compute a 5-component decomposed risk score. Read-only \u2014 does NOT persist or update Case.risk_band; the supervisor's intake fan-out owns that. Every invocation writes one ledger entry.",
        "operationId": "score_risk_v1_agents_risk_scoring_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskScoringInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskScore"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the Risk Scoring agent against a case",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/screening/run": {
      "post": {
        "description": "Calls the Screening agent (Story 6.2) with the supplied case_id and pre-built subject list (entity / directors / UBOs). Returns the typed output including auto-dismissed hits. Every invocation writes one ledger entry.",
        "operationId": "run_screening_v1_agents_screening_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreeningAgentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningAgentOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the Screening agent against a case's subjects",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/ubo_graph/build": {
      "post": {
        "description": "Calls the UBO Graph agent. Looks the CIN up via the mca_lookup tool, builds typed nodes + edges with the nominee heuristic, returns the graph. Every invocation writes one ledger entry.",
        "operationId": "build_ubo_graph_v1_agents_ubo_graph_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UBOGraphInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UBOGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build the UBO graph for a case",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/writing/draft": {
      "post": {
        "description": "Story 7.3 \u2014 invokes the Writing agent for a case in decision_ready (or pending_seal / committed for re-draft). Reads upstream agent outputs from the case's intake row, runs the LLM, and returns a typed DraftedRationale with Tiptap-renderable HTML and structured citations. Every invocation writes one ledger entry.",
        "operationId": "draft_rationale_v1_agents_writing_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritingAgentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftedRationale"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft the KYC decision rationale for a case",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}